+++ src/collective/cron/setuphandlers.py
@@ -1,12 +1,18 @@
 import logging
 from zope.component import getMultiAdapter
 from Products.CMFCore.utils import getToolByName
-from collective.cron import interfaces
+from collective.cron import interfaces as i
+from collective.cron import events as e
 from collective.cron import crontab
+from zope.component.hooks import getSite, setSite
+import transaction
 
 from StringIO import StringIO
+from zope.event import notify
 
 def setupVarious(context):
+    ### XXX making ._p_jar appear ###
+    transaction.savepoint()
 
     # Ordinarily, GenericSetup handlers check for the existence of XML files.
     # Here, we are not parsing an XML file, but we use this text file as a
@@ -21,14 +27,20 @@ def setupVarious(context):
     # initialiing manager will in turn intialize any
     # queue manager who is in charge for initializing
     # needed stuff for async jobs
+    oldsite = getSite()
+    setSite(portal)
     crt = crontab.Crontab.load()
     marker = getMultiAdapter(
         (portal, crt),
-        interfaces.ICrontabManager
+        i.ICrontabManager
     )
     crt.save()
 
+    setSite(oldsite)
+
 def setupCrons(context): # pragma: no cover
+    ### XXX making ._p_jar appear ###
+    transaction.savepoint()
     logger = logging.getLogger('collective.cron.exportCrons')
 
     # Ordinarily, GenericSetup handlers check for the existence of XML files.
@@ -48,12 +60,14 @@ def setupCrons(context): # pragma: no cover
     crt = crontab.Crontab.load()
     exportimporter = getMultiAdapter(
         (portal, crt),
-        interfaces.IExportImporter
+        i.IExportImporter
     )
     exportimporter.do_import(xml)
     logger.info("crontab imported")
 
 def exportCrons(context): # pragma: no cover
+    ### XXX making ._p_jar appear ###
+    transaction.savepoint()
     # if the queue is not marked, initialize the crontab !
     logger = logging.getLogger('collective.cron.exportCrons')
     portal = getToolByName(
@@ -65,7 +79,7 @@ def exportCrons(context): # pragma: no cover
         return
     exportimporter = getMultiAdapter(
         (portal, crt),
-        interfaces.IExportImporter
+        i.IExportImporter
     )
     result = exportimporter.do_export()
     context.writeDataFile("crons.xml", result, "txt/xml")