+++ src/collective/cron/upgrades/__init__.py
@@ -2,6 +2,10 @@
 
 import os, sys
 import logging
+import Zope2
+from zope.site.hooks import getSite, setSite
+
+from zope.component import getUtility, getMultiAdapter
 
 try:
     from Products.CMFPlone.migrations import migration_util
@@ -9,6 +13,7 @@ except:
     #plone4
     from plone.app.upgrade import utils as migration_util
 
+from AccessControl.SecurityManagement import getSecurityManager, setSecurityManager
 from Products.CMFCore.utils import getToolByName
 from Products.ATContentTypes.interface.image import IATImage
 from Products.ATContentTypes.content.image import ATImage
@@ -17,6 +22,7 @@ import transaction
 
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import _createObjectByType
+from collective.cron import interfaces as i
 
 PROFILE =  'collective.cron:default'
 PROFILEID = 'profile-%s' % PROFILE
@@ -93,4 +99,35 @@ def upgrade_1000(context): # pragma: no cover
     #portal_setup.runImportStepFromProfile('profile-collective.cron:default', 'propertiestool', run_dependencies=False)
     log('v1000 applied')
 
+def upgrade_2001(context): # pragma: no cover
+    """
+    """
+    site = getToolByName(context, 'portal_url').getPortalObject()
+    portal_setup = site.portal_setup
+    log = logging.getLogger(
+        'collective.cron.upgrade.2001')
+    try:
+        from plone.app.async.interfaces import IAsyncDatabase
+        from plone.app.async.service import AsyncService
+    except:
+        return
+    service = AsyncService()
+    db = service._db = getUtility(IAsyncDatabase)
+    asyncfs = db.databases.get(db.database_name)
+    itransaction = Zope2.zpublisher_transactions_manager
+    # see plone.app.async.service.AsyncService.getQueues
+    try:
+        service._conn = asyncfs.open()
+        service._conn.onCloseCallback(service.__init__)
+        queues = service.getQueues()
+        queue = queues['']
+        s = getMultiAdapter((site, queue), i.ICrontabMarker)
+        for idx, item in enumerate(s.annotations['plone'][:]):
+            if isinstance(item, tuple):
+                s.annotations['plone'][idx] = item[1]
+        itransaction.commit()
+    finally:
+        asyncfs.close()
+    setSite(site)
+    log.info('upgrade runned')